requirements.txt
Flask
pandas
openpyxl
requests
python-dotenv
playwright
